set(res)uid: Warn on unused result.
Not checking the result of setuid(2) is a known security bug: setuid(2)
can be made to fail by for example creating a lot of processes with the
target UID, at which point setuid(2) will fail with EAGAIN and the
process will still run as the old UID.

This is currently a problem for example in:
https://cs.corp.google.com/android/external/ipsec-tools/main.c?l=179

Try to make this better by warning on unused result. Many packages in
external/ don't compile with -Werror, but this is a start.

Bug: 34251640
Test: HiKey compiles with https://android-review.googlesource.com/#/c/321988/.
Test: Will check some internal targets before marking Verified+1.
Change-Id: If68ecfb6070b3fea4a130690b7ce1f7dbdd31994