Update patch set 1
Patch Set 1:

(2 comments)

Do we have any idea about the impact of this change on apps? There are plenty of apps that delegate to executables (ssh apps in particular do this often iirc, but it's bound to be elsewhere). Does this only affect armeabi stuff? Should have no impact on armeabi-v7a? We were talking about that in the office here and the best we could say was "I think so".

Assuming the minimal impact of the above, presumably this is something we'd need in the NDK if we want people to be able to use armeabi executables (at the very least this kills all current unit tests, right?)

Patch-set: 1
Label: Verified=0