Update patch set 1
Patch Set 1: Verified+1 Presubmit-Ready+1

This change is identical to what was merged in
https://android-review.googlesource.com/q/owner:self+status:merged+topic:prop_ctx_split
and then later reverted. This is the only patch that changed from that series
in order to fix b/34370523.

The problem was introduced when the new updater tries to find new (renamed) property_context files in bionic, failing recovery update.

I've verified this using 'adb sideload' on angler to make sure it can take the update that has these changes.

Patch-set: 1
Label: Verified=+1