Update patch set 1
Patch Set 1:

Is there any way that some kind of test can be written for this? We got burned before when crashing didn't work properly. If there isn't a test, then perhaps a test executable that is only built for userdebug and that will generate a syscomp failure. Sort of like our crasher app in system/core/debuggerd.

Patch-set: 1