Update patch set 2
Patch Set 2:

It would be good to bump up the thread stress test, however, I think you could write a dedicated test for this.

If you had a test that creates a thread, then find all of the maps associated with that thread. Now let the thread go away and verify that the maps are gone. You can do this with a join'd thread and a detached thread. Especially, since you are marking the maps, it should be possible to find the maps.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>