Deal with situation when pid > 65536 on 32 build.
This almost would never happen on a real android device if we don't
tune the default value of /proc/sys/kernel/pid_max. But if somebody
tuned it for some reason, the old code will just dead lock.
This isn't a perfect fix. But in practice, it should work: we aren't
expecting one process will spawn more than 64k threads and get the pid
wrapped.

Test: manual - boot emulator with this change.
Change-Id: If8f848bed538638d850eb412db4639b59a9de17e
Signed-off-by: Tao Wu <lepton@google.com>