Update patch set 5
Patch Set 5:

Nextbit Robin (msm8992) is a big-little device, but you have a single set of numbers.

You need to get numbers for each type of core to make sure that if there is an improvement on the big cores, that it doesn't cause performance to drop on the little cores. That's not easy with the bionic benchmarks, so you might try to use micro_bench --lock_to_cpu as a first test to see if there is anything really bad going on.

My test on a Pixel shows that this is a pretty bad regression, as much as ten times slower in the worst case. I'll test on my 6P (which is also a big-little) to see if that shows improvements.

At the very least this probably needs to be targeted a bit better.

Patch-set: 5
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>