Update patch set 5
Patch Set 5:

My quick check on a big-little device I have shows that this is an improvement for the big core, but a relatively large regression on the little core.

This seems like it would do better as a cortex-a53 only change. I also would not use this on big-little device since those devices will almost always be running on the little cores, so this would result in an overall regression of strlen on those systems.

Patch-set: 5
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>