Update patch set 1
Patch Set 1:

Hmm... that is surprising. I was under the impression that recent clangs have support for all of the builtins. 

Jake, I suggest submitting Clang bugs for any builtin that is not working at it should (or put otherwise - does not generate the exact code that you are trying to replace).

As for libm in general - aren't we missing an -ffreestanding ?

I was always under the impression that without this, on architecture that do not have the builtin support for certain operations and delegate it to the runtime we might end un in an infinite loop if:

libm_function() {
 return __builtin_that_calls_runtime_libm_function();
}

However, I need to spend more time understanding if freestanding is exactly what we need.

Patch-set: 1
Reviewer: Gerrit User 1072549 <1072549@85c56323-6fa9-3386-8a01-6480fb634889>