Update patch set 1
Patch Set 1: Code-Review+2

Feel free to ignore most of the comments then. But make_path returning -1 and no checking of make_path is really bad and should probably be fixed, but if you want to do something in a follow-on CL, that's fine with me.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2