Update patch set 3
Patch Set 3:

> That would be versioner

so this is blocked on us getting a new clang, and then jmgao switching over so we're not using #if any more. which would require us to abandon GCC? (do any of the legacy vendor projects still building with GCC use these versioned "NDK" headers?)

i guess the only alternative would be to work out which -- if any -- of the tree's NDK-header-using projects genuinely need to be older than 21 and moving them all forward (if possible)?

or we come up with a way to say "hey, versioner, leave this line alone"?

Patch-set: 3
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0