Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > That would be versioner
> 
> so this is blocked on us getting a new clang, and then jmgao switching over so we're not using #if any more. which would require us to abandon GCC? (do any of the legacy vendor projects still building with GCC use these versioned "NDK" headers?)
>
> i guess the only alternative would be to work out which -- if any -- of the tree's NDK-header-using projects genuinely need to be older than 21 and moving them all forward (if possible)?

It's not just the platform that matters here. If affects all NDK users.

> or we come up with a way to say "hey, versioner, leave this line alone"?

Or just remove the `__INTRODUCED_IN` tag from the real header for the time being. Technically the symbol does exist in all API levels after this change. It might be `NULL`, but it is defined.

Patch-set: 3
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>