Update patch set 4
Patch Set 4:

Yeah, I'm not too happy with all of the static_libs everywhere.

I think my new plan is to create the library as before, but I'm going to include the async source in libc directory, rather than as a library. I'll add an include path in the defaults, and I think it should look much better.

I'll also rename to async_safe_ for the macros.

I still plan on leaving the library as async_safe_log because I also cannot think of something else that will be in it.

Patch-set: 4
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>