Fix LTO .symver visibility
LLVM LTO sets all .symver versioned symbols to the visibility of the
target symbol. To work around this, all _impl{2} symbols must be global
for LTO to correctly make the symbol versions set with .symver
global. Alternatively we can try to fix this upstream, since the
versioned symbol is already set global in the linker maps so the
compiler should make the versioned symbol global regardless of the
visibility of the original. See https://reviews.llvm.org/D30485 and
https://bugs.llvm.org//show_bug.cgi?id=25679.

Test: built libc with LTO

Change-Id: I152ff7b0949263e87000bbd2efa136983ffec04f