libdl: Move android_get_exported_namespace()
This commit moves android_get_exported_namespace() from LIBC_PLATFORM
to LIBC.  We need this because LIBC_PLATFORM is completely ignored
when the LL-NDK stub library (for VNDK) was generated.

Nothing change to the NDK stub libraries because the symbols tagged
with #vndk is only visible when we are building LL-NDK stub libraries.

Test: Build aosp_arm-userdebug with BOARD_VNDK_VERSION := current
Change-Id: Ic1037fe7734ce8c1132a202dca1b0a7399357004