Update patch set 11
Patch Set 11:

(2 comments)

You're going to totally hate me (perhaps Elliot will as well), but could we split this up into a sequence of smaller changes for the sake of my pea brain?

I propose:

    [1] add qhook to netcontext, make definitions visible to netd

    [2a] gethostbyaddr/bionic
    [2b] gethostbyaddr/netd

    [3a] gethostbyname/bionic
    [3b] gethostbyname/netd

    [4] res_setnetcontext

2a&b should be submitted as a single topic
3a&b should be submitted as a (different) single topic

I feel that by the time we reach step 4 we'll know for sure when and where app_*/dns_* need to be passed and used.

I apologize.  It's just that this stuff is super-tricksy and we have an open bug (b/37347238) where I think what's happening is we're getting app and dns netid/mark wrong somewhere.

If you don't want to do this, I would be happy to take these changes and do the needful myself, no worries.

Patch-set: 11
Reviewer: Gerrit User 1000451 <1000451@85c56323-6fa9-3386-8a01-6480fb634889>