Fix crash during system properties tests/benchmarks
prop_info structures returned from the system properties APIs are
meant to be valid for the entire lifetime of a program, however the
system properties tests and benchmarks were invalidating these when
they created a temporary set of system property areas for testing.

This change creates two internal APIs for storing and restoring the
set of system property areas to maintain the validity of the prop_info
structures.

Bug: 62197783
Test: bionic-unit-tests, bionic-benchmarks

Change-Id: Ie4cbd2e6f73e884ecd91ca977eb022a81cfb4c46