Build support for 32-bit armv8-a
The assembly in arm's generic strlen implementation contains deprecated
instruction sequence:
    it      eq  ; start of IT block
    ldreq   ... ; 32-bit T32 insruction in IT block deprecated in armv8
This will cause compiler error because of -Winline-asm and -Werror.

The fix here is to change the sequence:
    it      eq
    ldreq   ...
    bne     1f
to equivalent sequence:
    bne     1f
    ldr     ...
The resulted sequence is (1 instruction) shorter.

See ARM for ARMv8 for details:
F6.2 Partial Deprecation of IT
... All uses of IT that apply to instructions other than a single
subsequent 16-bit instruction from a restricted set are deprecated, ...

Bug: 62895439
Test: "bionic-unit-tests-static --gtest_filter=*strlen*" on armv7
      emulator and armv8 sailfish.
      The former is built with armv8-a as its TARGET_2ND_CPU_ARCH;
      the latter is built with armv7-a as its TARGET_CPU_ARCH;
      Both TARGET(_2ND)_CPU_VARIANTs are set to "generic".

Change-Id: Ia2208b4e2ba2cad589975bc7b4617688cbb8826a