Make stdio buildable by the NDK.
No functional change intended. This just rearranges some pieces of
stdio so we can build them with the NDK (as part of
libandroid_support).

The definition of struct __sFILE is moved to stdio/bits/struct_file.h.
When building for older API levels, stdio.h includes the public
version of this header and gets a definition which conflicts with
ours. Match the public header name so we can shadow it.

The definition of fpos_t moves into bits/fpos_t.h to accommodate for
the above.

Move __sglue into its own source file. We don't need the rest of
stdio.cpp (and don't want duplicates of things like __sF).

Move the *wprintf family into their own source file for the same
reason. We may end up doing this with more functions, but for now all
libandroid_support needs is the *wprintf family.

Test: make checkbuild
Test: Able to add vswprintf.c and wprintf.cpp (and everything they
      need) to libandroid_support and pass libc++ tests in the NDK.
Bug: https://github.com/android-ndk/ndk/issues/300

Change-Id: I4127071906fc9dacf40e41cff35e403b48ba297c