Update patch set 2
Patch Set 2:

> but __cfi_slowpath should be kept together with __cfi_slowpath_diag.

Ok. I put them together.

> Either make it public ...

I don't want to make the symbol public to everybody. No vendor code is expected to depend on it directly. The symbol is needed just because the clang is assuming the existence of the symbols.

> ... or move symbol to non-ndk library.
Do you mean creating /vendor/lib/libdl_cfi.so and letting vendors to link against to it? I think this is worse than exposing the symbol to vendors because, IIUC, libdl_cfi.cpp is dependent on internals of libc (private/CFIShadow.h, etc.). If we make libdl_cfi.so and install it to vendor partition, then we are exposing the libc internal to vendors and we won't be able to modify that part of libc.

I think I don't clearly understand your concern. Could you tell me what can go wrong if I mark the symbols cfi_slowpath* as "# vndk"?

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>