Update patch set 1
Patch Set 1: Code-Review+1

This is to make it possible to load vendor libraries into the default namespace in a vendor process.

Contrary to system process where default namespace is for platform-only libs, default namespace in a vendor process will be for vendor libs (because main executable is in vendor) and platform-only libs will be loaded into the new 'system' namespace. 

So, when a platform-only lib (libvndksupport.so) is loading a vendor lib in a vendor process, it should  be able to get a reference of the default namespace so that it can load the vendor lib from there.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1