Update patch set 3
Patch Set 3:

> (1 comment)

The problem here is that you are using 26 which is already frozen.

Which means that after next ndk release apps compiled with target sdk version 25 will get invalid version for these symbols if they use them. It will result in apps crashing on OC.. Note that in OC these symbols are LIBC_PLATFORM. and this change will not affect it.

My understanding is that it should be either =future or whatever number we are using for target release if it is already know.

Patch-set: 3
Reviewer: Gerrit User 1041673 <1041673@85c56323-6fa9-3386-8a01-6480fb634889>