linker: libc.so is automatically loaded before libdl.so.
This is to handle the exceptional case of libdl.so: it is using the
symbol __cxa_finalize which is defined in libc.so but libdl.so does
not have libc.so in its DT_NEEDED entries. This is to prevent the
circular dependency when build the two libraries. However this is
causing link error when libdl.so is the first library that is being
loaded in a linked namespace other than the default namespace. In
that case, since libc.so is not loaded in the namespace, the linker
can't resolve the __cxa_finalize symbol. To solve the exception case,
libc.so is explicitly added to the load_tasks when the current lib
being loaded is libdl.so so that libc.so can be always loaded with
libdl.so.

Bug: 62815515
Test: boot and check if "CANNOT LINK EXECUTABLE" error shows after
      vendor process have additional namespace for LL-NDK.

Change-Id: I4ce1e0ed3041225927b0af8fdba1b9d805fe480c