libc: Split FORTIFY into its own headers
(Incomplete patch: once we're happy with the naming/placement of
everything, I'll do this to the other 7 FORTIFY headers. Commit message
is "final", though.)

This patch cleans up our standard headers by moving all of the FORTIFY
cruft out in to its own sandbox. In order to include the *_chk and
*_real declarations, you can either enable FORTIFY, or `#define
__BIONIC_DECLARE_FORTIFY_HELPERS`.

This also makes the versioner ignore the FORTIFY implementation
headers, since we're guaranteed to pick the FORTIFY'ed headers up when
looking at the regular headers. (...Not to mention that making the
FORTIFY'ed headers freestanding would be annoying to do and maintain
for ~no benefit).

This patch bakes the knowledge of where FORTIFY headers live directly
into the versioner. We could go with a more general approach (e.g.
adding an -X IGNORED_FILE flag that tells the versioner to ignore
$HEADER_PATH/$IGNORED_FILE), but we'd then have to repeat that for
every test, every manual invocation of the versioner, etc. for no
benefit that's obvious to me.

Bug: 12231437
Test: m checkbuild on bullhead internal master + CtsBionicTestCases. no
new errors.

Change-Id: Iffc0cc609009b33d989cdaddde0a809282131a5b