Update patch set 2
Patch Set 2:

The problem is that this code is not used by any currently tested config. So even if you make this change, we don't compile it by default, and we don't even know if it works.

At some point, we'd like to get rid of this file completely.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>