Update patch set 1
Patch Set 1:

> > (1 comment)
 > >
 > > not clear to me whether this is a clang bug or WAI. i certainly
 > > realize this is very unusual code.
 > 
 > I don't think this is a bug, as a "static" function that isn't
 > explicitly used should be optimized out. Is _start_main explicitly
 > specified in the C standard, or is this a posix or ELF thing? I
 > would expect Clang to treat this function no differently than any
 > other user-defined function. From a brief look around Clang
 > sources, there is no special handling for functions named
 > "_start_main".

no, it's a name i made up. the only thing to keep it alive is the fact that it's referenced by the inline assembler's b/call instructions. it seems like gcc/gas (not sure which) does pay enough attention to the inline assembler for that to work, but clang doesn't.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0