Add ushort/uint/ulong typedefs.
We already had the u_short/u_int/u_long typedefs, but util-linux uses the
variants without the underscores, which BSD and glibc both also offer. BSD
also has an `unchar` equivalent to u_char in the same "SysV compatibility"
section, but glibc doesn't so it's unlikely to be widely used.

Bug: N/A
Test: built util-linux with a modified standalone toolchain
Change-Id: I7a83cc1c1025f7de5c12e6ef350a32fb2cb3442b