libc fortify: warn on open() with useless mode bits
This also expands our warnings/errors to consider O_TMPFILE. I think
it's reasonable to not care about mode bits on a temporary file, but I
dislike the seemingly-undefinedness of just leaving the last arg off.

If there's some magic we do to guarantee that the last arg is always 0,
I'm happy to no longer complain about O_TMPFILE.

Bug: 64132680
Test: checkbuild on bullhead internal master; no new CtsBionicTestCases
fail.

Change-Id: I98762bbee0ecdd7b4ffc21135992f2eed320317f