Warn on bad usages of pthread_cond_timedwait()
If pthread_cond_timedwait() is called on a condition variable that is
using CLOCK_REALTIME, then when the system time changes while a
process is waiting on the condition variable it may time out early or
later than expected.  The latter is open ended; if a user changes the
time from 2010 to 2015, then the condition variable will essentially
never time out.

This is compounded by the fact that pthread condition variables'
default initialization is CLOCK_REALTIME.

There are 2 possible solutions at this time for users to fix this
issue:
1) Opt into CLOCK_MONOTONIC when the condition variable is
initialized, see
https://android-review.googlesource.com/#/c/platform/system/core/+/342638/
as an example of how to do this.
2) Use the Condition class from libutils.

Note that std::condition_variable is also broken in this way with no
fix in sight (b/35756266)

Bug: 64623895
Bug: 64694413
Test: boot bullhead
Change-Id: Icfce5ef45c4688e04c60be60ab08e8b2f0e1b9c0