Always use CLOCK_MONOTONIC for pthread_cond_timedwait()
pthread condition variables' default initialization is to use
CLOCK_REALTIME, however this behavior is essentially never intended,
as that clock is prone to change discontinuously.  What users really
intend is to use CLOCK_MONOTONIC, however there is a large amount of
existing code that does not opt into CLOCK_MONOTONIC, and we have seen
numerous bugs directly attributable to this difference.

Therefore, this change provides a general workaround to always use
CLOCK_MONOTONIC for waiting, regardless of what the input timespec
is.

Bug: 64694413
Bug: 64623895
Bug: 35756266
Test: boot bullhead, sailfish
Test: observe condition variables + clock changes
Test: bionic pthread unit tests
Change-Id: Ie0c303b779807426f40459615e320be4d42cc147