Update patch set 3
Patch Set 3:

(7 comments)

> especially if you want to change other wait functions like pthread_mutex_timedlock()?

I've thought about this too.  We haven't seen any issues that are directly related to this, but I'm sure the same overall issue is present there (aka why would anyone _ever_ want a timedlock() with CLOCK_REALTIME).  I'm guessing we're not hitting this since there are many fewer users of pthread_mutex_timedlock() than pthread_cond_timedwait().

I could be easily convinced to add this preemptively to all of the timed waits in pthreads.  Looked like pthread_rwlock_timedrdlock() and pthread_rwlock_timedwrlock() are the only other two we're missing.  

Then there's sem_timedwait() too.  I don't see any users of this in our tree that are actually intending to use CLOCK_REALTIME.  A vast majority simply add x seconds to current time.  A handful even wrap the call in a sem_timed_relative_wait() type function.

Given that, we could just move this to __futex_wait_ex().  It wouldn't be compliant behavior, but it'd be the same behavior that Android had before the move to FUTEX_WAIT_BITSET, and frankly, it'd be the _correct_ behavior for all of the APIs built on top of this.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>