Refactor send data.
The old code would always return the number of bytes that would
have been written into the buffer. However, everyone used it as
an indication of the number of bytes actually written. Modify it to
conform to this expectation.

Add tests to verify this new behaviour.

Also, implement async_safe_format_buffer to call
async_safe_format_buffer_va_list.

Test: Ran new unit tests, booted on angler.
Change-Id: I7fb13e209f5b7443d212f55aab4b05ff2e0e8219