Update patch set 1
Patch Set 1:

I have talked to the Aarch64 maintainer for the kernel, and according to him, that has been the behaviour for a long time. The kernel developers treat it as ABI at this point, so it is highly unlikely to change. In any case, the kernel has to sanitize the register values to defend against any information leaks to user space, so it might as well preserve the registers.

I have also examined the 4.x kernel source, and in particular the kernel_entry and kernel_exit macros. However, I agree with the point about adding tests, so I will work on that as soon as possible.

Patch-set: 1
Reviewer: Gerrit User 1084448 <1084448@85c56323-6fa9-3386-8a01-6480fb634889>