Update patch set 2
Patch Set 2: Code-Review-1

Yep actually I was already on this
https://android-review.googlesource.com/c/platform/bionic/+/492818

Can I stop suspending commiting this and move forward with my version?
The ChromeOS device is on N-MR1 hence the test for N-MR1 must be fixed. (and the commit to nougat-mr1-cts-dev will be automerged to the master anyway.

Patch-set: 2
Reviewer: Gerrit User 1089769 <1089769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1